+++ debug_toolbar/utils.py
@@ -14,12 +14,6 @@ from django.utils.safestring import SafeString, mark_safe
 
 from debug_toolbar import _stubs as stubs, settings as dt_settings
 
-try:
-    import threading
-except ImportError:
-    threading = None
-
-
 _local_data = Local()
 
 
@@ -357,33 +351,3 @@ def get_stack_trace(*, skip=0):
 def clear_stack_trace_caches():
     if hasattr(_local_data, "stack_trace_recorder"):
         del _local_data.stack_trace_recorder
-
-
-class ThreadCollector:
-    def __init__(self):
-        if threading is None:
-            raise NotImplementedError(
-                "threading module is not available, "
-                "this panel cannot be used without it"
-            )
-        self.collections = {}  # a dictionary that maps threads to collections
-
-    def get_collection(self, thread=None):
-        """
-        Returns a list of collected items for the provided thread, of if none
-        is provided, returns a list for the current thread.
-        """
-        if thread is None:
-            thread = threading.current_thread()
-        if thread not in self.collections:
-            self.collections[thread] = []
-        return self.collections[thread]
-
-    def clear_collection(self, thread=None):
-        if thread is None:
-            thread = threading.current_thread()
-        if thread in self.collections:
-            del self.collections[thread]
-
-    def collect(self, item, thread=None):
-        self.get_collection(thread).append(item)